requirements.txt
streamlit
opencv-python-headless
mediapipe
numpy
scikit-learn
Pillow